Add Cart Functionality
So we can add products to a cart now. Carts are searched by the
current_users id, rather than cart_id, so it's not truly RESTful but w/e
I have it set up currently that if a person isn't logged in, the cart
button links to login, else it will show cart

Adding products to cart is done remotely, but I want to add Jquery to
display a notice so it isn't so silent.

Still need to implement checkout and remove/update items in cart